+++ packages/tools/src/union.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 /**
  * @since 0.0.1
  * @template T The type to get the union from
@@ -15,4 +14,9 @@ export type UnionFromTuple<T> = T extends (infer U)[] ? U : never;
  * type Result = IntersectionFromTuple<['foo', 'bar', 1]>
  * // Result = 'foo' & 'bar' & 1
  */
-export type UnionToIntersection<U> = (U extends any ? (k: U) => void : never) extends (k: infer I) => void ? I : never;
+export type UnionToIntersection<U> = (
+  U extends unknown ? (k: U) => void : never
+) extends (k: infer I) => void
+  // dprint-ignore
+  ? I
+  : never;